Completes coupon code functionalities

+++ app/controllers/profile/orders_controller.rb
@@ -8,6 +8,9 @@ class Profile::OrdersController < ApplicationController
 
   def show
     @order = Order.find(params[:id])
+    if @order.coupon_id != nil
+      @coupon = Coupon.find(@order.coupon_id)
+    end
   end
 
   def destroy
@@ -31,12 +34,37 @@ class Profile::OrdersController < ApplicationController
   end
 
   def create
-    order = Order.create(user: current_user, status: :pending)
-    cart.items.each do |item, quantity|
-      order.order_items.create(item: item, quantity: quantity, price: item.price)
+    if coupon != nil && coupon.valid_for_user?(current_user) == false
+      flash[:danger] = "You have already used coupon #{coupon.name}."
+      session[:coupon_code].clear
+      redirect_to cart_path
+    else
+      if coupon == nil
+        order = Order.create(user: current_user, status: :pending)
+      else
+        order = Order.create(user: current_user, status: :pending, coupon_id: coupon.id)
+        coupon_value = coupon.value
+      end
+      cart.items.each do |item, quantity|
+        if percent_coupon_applied?(item)
+          item_price = coupon.discount_percent_off(item)
+          order.order_items.create(item: item, quantity: quantity, price: item_price)
+        elsif dollar_coupon_applied?(item)
+          item_price = coupon.discount_dollar_off(item, quantity, coupon_value)
+          coupon_value = 0
+          if item_price.class == Array
+            price = item_price[0]
+            coupon_value = item_price[1]
+          end
+          order.order_items.create(item: item, quantity: quantity, price: item_price)
+        else
+          order.order_items.create(item: item, quantity: quantity, price: item.price)
+        end
+      end
+      session.delete(:cart)
+      session.delete(:coupon_code)
+      flash[:success] = "Your order has been created!"
+      redirect_to profile_orders_path(@discounted_total)
     end
-    session.delete(:cart)
-    flash[:success] = "Your order has been created!"
-    redirect_to profile_orders_path
   end
 end
